chore: BIL-89: moved query boundaries related to source creation and finding to base boundary so all components have access to it
