+++ services/database/boundary/base_boundary.py
@@ -2,6 +2,8 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from manager.dbmanager import DBManager
 
+import json
+
 class QueryBoundary:
     def __init__(self, db_manager: "DBManager"):
         self.db = db_manager
@@ -62,6 +64,15 @@ class ReadBoundary(QueryBoundary):
         """
         result = self.db.fetch_all(query)
         return result
+    
+    def find_source(self, 
+            code: str
+        ) -> int:
+        query = """
+            SELECT id FROM audit.sources WHERE code=%s
+        """
+        source_id = self.db.fetch_clean_one(query, (code,))
+        return source_id
 
 class WriteBoundary(QueryBoundary):
     def persist_label_studio_project(self,
@@ -112,6 +123,64 @@ class WriteBoundary(QueryBoundary):
         """
         file_id = self.db.fetch_clean_one(query, (etag, type, object_path, bucket, source_id, version_id, version_note, data_format, content_hash))
         return file_id
+    
+    def init_source(self,
+            source_type: str,
+            url: str
+        ):
+        query = """
+            INSERT INTO audit.sources(source_type, url)
+            VALUES (%s, %s)
+            RETURNING id;
+        """
+        source_id = self.db.execute(query, (source_type, url))
+        return source_id
+    
+    def update_source(self,
+            code: str,
+            name: str,
+            description: str,
+            version: str, 
+            note: str,
+            parent_source: str,
+            official_citation: str = None,
+            date_published: str = None,
+            metadata: json = None
+        ):
+        query = """
+            UPDATE audit.sources
+            SET code = %s,
+                name = %s,
+                description = %s,
+                version = %s,
+                note = %s,
+                official_citation = %s,
+                date_published = %s,
+                metadata = %s,
+            WHERE id = %s
+        """
+        self.db.fetch_clean_one(query, (code, name, description, version, note, parent_source, official_citation, date_published, metadata))
+
+    def persist_source(self,
+            source_type: str,
+            code: str,
+            name: str,
+            description: str,
+            version: str,
+            url: str,    
+            note: str,
+            parent_source: str,
+            official_citation: str = None,
+            date_published: str = None,
+            metadata: json = None
+        ) -> int:
+        query = """
+            INSERT INTO audit.sources (source_type, code, name, description, version, url, note, parent_source, official_citation, date_published, metadata) 
+            VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
+            RETURNING id;
+        """
+        source_id = self.db.fetch_clean_one(query, (source_type, code, name, description, version, url, note, parent_source, official_citation, date_published, metadata))
+        return source_id
     
 class DeleteBoundary(QueryBoundary):
     pass
